createFineGrainedBarChart -> FineGrainedBarChart
